+++ src/main/java/com/example/palayo/domain/auth/dto/request/LoginUserRequestDto.java
@@ -1,4 +1,4 @@
-package com.example.palayo.auth.dto.request;
+package com.example.palayo.domain.auth.dto.request;
 
 import jakarta.validation.constraints.Email;
 import jakarta.validation.constraints.NotBlank;
@@ -6,10 +6,12 @@ import jakarta.validation.constraints.Pattern;
 import lombok.AllArgsConstructor;
 import lombok.Getter;
 
+//TODO Class 명은 카멜케이스입니다
 @Getter
 @AllArgsConstructor
-public class loginUserRequestDto {
+public class LoginUserRequestDto {
     private Long id;
+
     @NotBlank
     @Email(message = "이메일 형식에 어긋납니다")
     private String email;
