+++ src/main/java/com/example/palayo/domain/user/entity/PointHistories.java
@@ -1,4 +1,4 @@
-package com.example.palayo.domain.user.entity.log;
+package com.example.palayo.domain.user.entity;
 
 import com.example.palayo.domain.user.entity.User;
 import com.example.palayo.domain.user.enums.PaymentType;
@@ -8,6 +8,7 @@ import org.springframework.data.annotation.CreatedDate;
 
 import java.time.LocalDateTime;
 
+//TODO 굳이 log로 나눌 필요 없습니다. Depth만 늘어나요
 @Entity
 @Getter
 public class PointHistories {
